+++ packages/gov-ui-kit/src/components/index.ts
@@ -10,3 +10,4 @@ export * from './progress';
 export * from './spinner';
 export * from './switch';
 export * from './tag';
+export * from './toggles';